refactor mobilenavigation to separate components and close sidebar when user clicks a link
